refactor(v3): extract bridge JavaScript into real asset files
The ~1,950 lines of JavaScript embedded in a Dart string inside
generateIndexHtml move byte-for-byte into assets/monaco/bridge/*.js
(core, focus, viewport-fit, editor-api, scroll-handoff, lsp, boot).
Extraction used the 2.3.0 generator output as ground truth; only the
installer glue (window.__FMB wrappers with a shared ctx) is new, and
boot.js preserves the original execution order. The current protocol
(window.flutterMonaco, flutterMonacoInvoke/Async) is unchanged.

- lib/src/assets/html_builder.dart: slim page template (CSP, styles,
  platform shims, window.__FM_PAGE, script tags); generateIndexHtml
  delegates to it and gains bridgeBasePath
- htmlGenerationVersion and the HTML caches are deleted: bridge files
  are rewritten on every launch and per-config HTML on every load, so
  stale-bridge bugs are impossible instead of versioned around
- pubspec ships assets/monaco/bridge/; native extraction always
  refreshes it (sentinel still guards only min/)
- tests that scraped generateIndexHtml output now read the bridge
  sources; new structural test pins file existence, load order, and
  installer markers

Verified: flutter analyze clean, 446 tests green, example app boots
with identical console flow on macOS (WKWebView, file://) and Chrome
(blob iframe, resolved asset URLs).
