+++ lib/src/core/monaco_assets.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
+import 'package:flutter_monaco/src/assets/html_builder.dart';
 import 'package:flutter_monaco/src/core/monaco_asset_storage.dart';
 
 /// Manages Monaco Editor assets across all platforms.
@@ -54,52 +55,7 @@ class MonacoAssets {
   /// native platforms to ensure the correct version is used.
   static const String monacoVersion = '0.55.1';
 
-  /// Cache-busting version for generated HTML and the JS bridge contract.
-  ///
-  /// Native platforms cache generated `monaco_<key>.html` files inside the
-  /// per-[monacoVersion] directory. When the bundled Monaco version doesn't
-  /// change but the generated HTML or `window.flutterMonaco` bridge shape
-  /// does, callers must include this constant in their cache key so stale
-  /// HTML from prior package versions is regenerated on first load.
-  ///
-  /// Bump this whenever [generateIndexHtml] output or the JS bridge changes
-  /// in a way Dart depends on.
-  ///
-  /// v5: added the LSP bridge (`flutterMonaco.lsp`, `flutterMonacoInvokeAsync`)
-  /// and the edge scroll handoff module (`flutterMonaco.setScrollHandoff` and
-  /// the `scrollHandoff` event).
-  static const int htmlGenerationVersion = 5;
-
   static Completer<void>? _initCompleter;
-
-  /// Validates and joins extra CSP `connect-src` source expressions.
-  ///
-  /// Returns either an empty string or a leading-space-joined list ready to
-  /// splice after `connect-src 'self' blob:`. Throws [ArgumentError] for
-  /// entries that could break out of the policy attribute or smuggle in
-  /// additional directives (whitespace, quotes, `;`, `<`, `>`).
-  static String _sanitizeConnectSources(List<String> sources) {
-    if (sources.isEmpty) return '';
-    final forbidden = RegExp(r'''[\s;'"<>]''');
-    final cleaned = <String>[];
-    for (final raw in sources) {
-      final source = raw.trim();
-      if (source.isEmpty) continue;
-      if (forbidden.hasMatch(source)) {
-        throw ArgumentError.value(
-          raw,
-          'allowedConnectSources',
-          'Each entry must be a single CSP source expression without '
-              'whitespace, quotes, or ";" (e.g. "wss://lsp.example.com")',
-        );
-      }
-      cleaned.add(source);
-    }
-    return cleaned.isEmpty ? '' : ' ${cleaned.join(' ')}';
-  }
-
-  // HTML cache to avoid regenerating the same HTML multiple times
-  static final Map<int, String> _htmlCache = {};
 
   /// Ensures Monaco assets are extracted and ready for use on native platforms.
   ///
@@ -145,6 +101,15 @@ class MonacoAssets {
         htmlFileName: _htmlFileName,
         monacoVersion: monacoVersion,
       );
+      // The sentinel guards only the large Monaco `min/` tree. The bridge
+      // JavaScript is rewritten on every launch so a package upgrade can
+      // never ship stale bridge code (replaces the 2.x
+      // `htmlGenerationVersion` cache-bust constant).
+      await refreshMonacoBridgeAssets(
+        assetBaseDir: assetBaseDir,
+        cacheSubDir: _cacheSubDir,
+        monacoVersion: monacoVersion,
+      );
       completer.complete();
     } catch (e, st) {
       _initCompleter = null;
@@ -182,10 +147,7 @@ class MonacoAssets {
 
     await ensureReady();
 
-    final cached = _htmlCache[cacheKey];
-    if (cached != null) return cached;
-
-    return _htmlCache[cacheKey] = await monacoAssetHtmlPath(
+    return monacoAssetHtmlPath(
       cacheKey: cacheKey,
       assetBaseDir: assetBaseDir,
       cacheSubDir: _cacheSubDir,
@@ -260,7 +222,6 @@ class MonacoAssets {
     // On web, only clear in-memory caches (no file system access)
     if (kIsWeb) {
       _initCompleter = null;
-      _htmlCache.clear();
       debugPrint('[MonacoAssets] Web cache cleared (in-memory only)');
       return;
     }
@@ -271,66 +232,28 @@ class MonacoAssets {
       monacoVersion: monacoVersion,
     );
 
-    // Reset the init completer and HTML cache
+    // Reset the init completer so ensureReady will re-extract.
     _initCompleter = null;
-    _htmlCache.clear();
   }
 
   /// Generates the HTML document that hosts the Monaco Editor.
   ///
-  /// This method creates a complete HTML page with:
-  /// - Monaco loader and editor initialization
-  /// - Platform-specific worker shims (blob URLs for WKWebView, etc.)
-  /// - JavaScript bridge (`window.flutterMonaco`) for Flutter communication
-  /// - Content-Security-Policy headers for security
-  /// - Optional custom CSS injection
+  /// Since 3.0 the page is a slim shell: platform channel/worker shims and a
+  /// `window.__FM_PAGE` config snippet stay inline, while all bridge logic is
+  /// loaded from real JavaScript files via `<script src>` (see
+  /// `assets/monaco/bridge/`). [bridgeBasePath] is the URL prefix for those
+  /// files without a trailing slash: the default `bridge` suits native
+  /// platforms where the generated HTML sits next to the extracted `bridge/`
+  /// directory; Windows passes an absolute `file://` directory URL and web
+  /// passes a resolved asset URL.
   ///
-  /// ### Parameters
-  ///
-  /// - [vsPath]: The path to the Monaco `vs/` directory. This varies by platform:
-  ///   - **Windows:** Absolute `file://` URL (e.g., `file:///C:/path/min/vs`)
-  ///   - **macOS/iOS:** Relative path (e.g., `min/vs`)
-  ///   - **Web:** Full URL resolved from base (e.g., `http://host/assets/.../min/vs`)
-  ///
-  /// - [isWindows]: Set `true` for Windows to configure WebView2 communication
-  ///   via `chrome.webview.postMessage`.
-  ///
-  /// - [isIosOrMacOS]: Set `true` for Apple platforms to enable blob URL worker
-  ///   shims required by WKWebView's `file://` restrictions.
-  ///
-  /// - [isWeb]: Set `true` for web platform to use iframe `postMessage` and
-  ///   dynamic loader script injection.
-  ///
-  /// - [customCss]: Optional CSS string injected into a `<style>` tag. Use for
-  ///   custom fonts (`@font-face`), theme overrides, or UI tweaks.
-  ///
-  /// - [allowCdnFonts]: If `true`, relaxes CSP to allow `https:` in `style-src`
-  ///   and `font-src`, enabling CDN-hosted fonts. **Security note:** This allows
-  ///   network requests from the editor.
-  ///
-  /// - [allowedConnectSources]: Additional CSP `connect-src` source
-  ///   expressions (e.g. `wss://lsp.example.com` or `ws://127.0.0.1:3000`).
-  ///   Required for WebSocket language servers - the default policy of
-  ///   `'self' blob:` blocks every `ws://`/`wss://` handshake. Each entry
-  ///   must be a single CSP source expression; entries containing
-  ///   whitespace, quotes, or `;` throw an [ArgumentError] to prevent
-  ///   policy injection. **Security note:** every listed origin becomes
-  ///   reachable from any JavaScript running inside the editor page.
-  ///
-  /// ### The `flutterMonaco` Bridge
-  ///
-  /// The generated HTML defines `window.flutterMonaco` with methods like:
-  /// - `getValue()`, `setValue(v)` - Content management
-  /// - `setTheme(t)`, `setLanguage(l)` - Editor configuration
-  /// - `getSelection()`, `setSelection(r)` - Selection handling
-  /// - `findMatches(q, opts)`, `replaceMatches(q, r, opts)` - Search
-  /// - `registerCompletionSource(cfg)` - Autocompletion
-  ///
-  /// Events are sent to Flutter via `flutterChannel.postMessage(json)`.
-  ///
-  /// See also:
-  /// - [MonacoController] which calls the `flutterMonaco` methods.
-  /// - The Monaco bridge, which receives events from the HTML.
+  /// All other parameters keep their 2.3.0 semantics: [vsPath] locates the
+  /// Monaco `vs/` directory, [isWindows]/[isIosOrMacOS]/[isWeb] select the
+  /// platform shims, [messageToken] authenticates web postMessage traffic,
+  /// [customCss] is injected into a `<style>` tag, [allowCdnFonts] relaxes
+  /// CSP for `https:` styles/fonts, and [allowedConnectSources] extends the
+  /// CSP `connect-src` list (validated; throws [ArgumentError] on entries
+  /// that could break the policy attribute).
   static String generateIndexHtml(
     String vsPath, {
     bool isWindows = false,
@@ -340,2079 +263,18 @@ class MonacoAssets {
     String? customCss,
     bool allowCdnFonts = false,
     List<String> allowedConnectSources = const [],
+    String bridgeBasePath = 'bridge',
   }) {
-    final extraConnectSources = _sanitizeConnectSources(allowedConnectSources);
-    // Platform-specific initialization scripts
-    String platformScript = '';
-
-    if (isWeb) {
-      platformScript =
-          '''
-<script>
-  console.log('[Web Init] Setting up for iframe mode');
-  self.MonacoEnvironment = {
-    baseUrl: '$vsPath/../',
-    getWorkerUrl: function(moduleId, label) {
-      var workerSrc = "self.MonacoEnvironment = { baseUrl: '$vsPath/../' }; importScripts('$vsPath/base/worker/workerMain.js');";
-      return URL.createObjectURL(new Blob([workerSrc], { type: 'application/javascript' }));
-    }
-  };
-  window.flutterChannel = {
-    postMessage: function(msg) {
-      window.parent.postMessage(msg, '*');
-    }
-  };
-  window.flutterMonacoToken = '${messageToken ?? ''}';
-  window.flutterMonacoPostMessage = function(message) {
-    var token = window.flutterMonacoToken;
-    if (typeof message !== 'string') {
-      if (token && message && typeof message === 'object') {
-        message._flutterToken = token;
-      }
-      message = JSON.stringify(message);
-    } else {
-      try {
-        var parsed = JSON.parse(message);
-        if (parsed && typeof parsed === 'object') {
-          if (token) parsed._flutterToken = token;
-          message = JSON.stringify(parsed);
-        }
-      } catch (_) {}
-    }
-    if (window.flutterChannel && window.flutterChannel.postMessage) {
-      window.flutterChannel.postMessage(message);
-    }
-  };
-  console.log('[Web Init] flutterChannel created successfully');
-</script>
-''';
-    } else if (isWindows) {
-      platformScript = '''
-<script>
-  // Windows: Create flutterChannel immediately when document is created
-  console.log('[Windows Init] Creating flutterChannel on document creation');
-  window.flutterChannel = {
-    postMessage: function(msg) {
-      if (window.chrome && window.chrome.webview) {
-        window.chrome.webview.postMessage(msg);
-      } else {
-        console.error('[flutterChannel] WebView2 API not available!');
-      }
-    }
-  };
-  console.log('[Windows Init] flutterChannel created successfully');
-</script>
-''';
-    } else if (isIosOrMacOS) {
-      // iOS and macOS need blob worker shim for WKWebView + file:// protocol
-      platformScript =
-          '''
-<script>
-  (function () {
-    console.log('[Init] Setting up worker shim for WKWebView');
-    const vsRel = '$vsPath'; // e.g., "min/vs"
-    let absVs;
-    try { absVs = new URL(vsRel, window.location.href).toString(); }
-    catch (_) { absVs = vsRel; }
-
-    // Ensure baseUrl points to the ".../min/" folder (not ".../min/vs")
-    const idx = absVs.lastIndexOf('/vs');
-    const baseUrl = idx >= 0 ? absVs.substring(0, idx + 1) : absVs; // e.g., ".../min/"
-
-    // Set baseUrl so Monaco can resolve URLs before workers start
-    self.MonacoEnvironment = {
-      baseUrl: baseUrl,
-      getWorkerUrl: function (moduleId, label) {
-        // Include the label for better worker resolution and debugging.
-        // This template is a cooked Dart string: the newline escapes must be
-        // written as \\n so JavaScript receives "\\n" and not a raw newline,
-        // which is a SyntaxError inside a JS string literal and would kill
-        // this whole script block (workers then fall back to the main thread).
-        const src =
-          "self.MonacoEnvironment = { baseUrl: '" + baseUrl + "' };\\n" +
-          "self.monacoLabel = '" + label + "';\\n" +
-          "importScripts('" + absVs + "/base/worker/workerMain.js');\\n";
-        return URL.createObjectURL(new Blob([src], { type: 'application/javascript' }));
-      }
-    };
-    console.log('[Init] Worker shim configured. baseUrl=' + baseUrl);
-  })();
-</script>
-''';
-    } else {
-      // Linux and other platforms: just set baseUrl
-      final baseUrl = vsPath.replaceAll('/vs', '/');
-      platformScript =
-          '''
-<script>
-  // Linux/Other: Set base URL for worker resolution
-  console.log('[Init] Setting Monaco base URL');
-  self.MonacoEnvironment = { baseUrl: '$baseUrl' };
-</script>
-''';
-    }
-
-    const jsEscapePattern = r'[.*+?^${}()|[\]\\]';
-
-    return '''
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
-    <meta
-      http-equiv="Content-Security-Policy"
-      content="default-src 'self' file: 'unsafe-inline' 'unsafe-eval'; script-src 'self' file: 'unsafe-inline' 'unsafe-eval'; style-src 'self' 'unsafe-inline'${allowCdnFonts ? ' https:' : ''}; font-src 'self' file: data:${allowCdnFonts ? ' https:' : ''}; img-src 'self' data: blob: file:; worker-src 'self' blob:; connect-src 'self' blob:$extraConnectSources;"
-    />
-    <!-- NOTE: connect-src intentionally limits in-page requests to self/blob.
-         Opt into remote endpoints (e.g. WebSocket language servers) via the
-         allowedConnectSources parameter instead of editing this policy. -->
-    <style>
-      html, body, #editor-container {
-        width: 100%; height: 100%; margin: 0; padding: 0; overflow: hidden;
-      }
-      ${isWeb ? '''
-      /* Scroll containment for the iframe embed (issue #11). Monaco only
-         preventDefault()s touches that start on its text surface or
-         scrollbar sliders; touches that start on the margin, a scrollbar
-         track, or before Monaco loads fall through to the browser's native
-         pan. Nothing in this document can scroll, so mobile Safari chains
-         that pan to the host Flutter page. Declaring the policy here removes
-         the native pan for every touch in this document; Monaco's own touch
-         scrolling is JS-driven and unaffected by touch-action. */
-      html, body, #editor-container {
-        touch-action: none;
-        overscroll-behavior: none;
-      }''' : ''}
-    </style>
-    ${customCss != null ? '<style id="flutter-monaco-custom">\n$customCss\n</style>' : ''}
-    $platformScript
-  </head>
-  <body>
-    <div id="editor-container"></div>
-
-    <script>
-      var require = { paths: { vs: '$vsPath' } };
-      console.log('[Monaco HTML] Require config set. VS_PATH is: ' + '$vsPath');
-    </script>
-
-    ${isWeb ? '''
-    <script>
-      // Web: Load loader.js dynamically to ensure proper timing in blob URL context
-      var loaderScript = document.createElement('script');
-      loaderScript.src = '$vsPath/loader.js';
-      loaderScript.onload = function() {
-        console.log('[Monaco HTML] loader.js dynamically loaded.');
-        window._monacoLoaderReady = true;
-        if (window._initMonacoWhenReady) window._initMonacoWhenReady();
-      };
-      loaderScript.onerror = function() {
-        console.error('[Monaco HTML] FATAL: loader.js FAILED TO LOAD.');
-        if (window.flutterMonacoPostMessage) {
-          window.flutterMonacoPostMessage({ event: 'error', message: 'Failed to load Monaco loader.js' });
-        } else if (window.flutterChannel) {
-          window.flutterChannel.postMessage(JSON.stringify({ event: 'error', message: 'Failed to load Monaco loader.js' }));
-        }
-      };
-      document.head.appendChild(loaderScript);
-    </script>
-    ''' : '''
-    <script src="$vsPath/loader.js"
-            onload="console.log('[Monaco HTML] loader.js successfully loaded.')"
-            onerror="console.error('[Monaco HTML] FATAL: loader.js FAILED TO LOAD.')"
-    ></script>
-    '''}
-
-    <script>
-      ${isWeb ? '''
-      // Web: Wait for loader.js to be ready
-      function _initMonaco() {
-        if (!window._monacoLoaderReady) {
-          window._initMonacoWhenReady = _initMonaco;
-          return;
-        }
-      ''' : '''
-      (function _initMonaco() {
-      '''}
-      console.log('[Monaco HTML] Attempting to require editor.main...');
-      try {
-        require(
-          ['vs/editor/editor.main'],
-          function () {
-            console.log('[Monaco] SUCCESS: editor.main.js has loaded. Initializing editor...');
-
-            function postMessageToFlutter(message) {
-              if (window.flutterMonacoPostMessage) {
-                window.flutterMonacoPostMessage(message);
-                return;
-              }
-              if (typeof message !== 'string') {
-                message = JSON.stringify(message);
-              }
-              if (window.flutterChannel && window.flutterChannel.postMessage) {
-                window.flutterChannel.postMessage(message);
-              } else {
-                console.error('[Monaco] Flutter communication channel is not available.');
-              }
-            }
-
-            monaco.editor.onDidCreateEditor(function (editor) {
-              window.editor = editor;
-
-              // Send live statistics updates
-              const sendStats = () => {
-                if (!editor.getModel() || !editor.getSelection()) return;
-                const model = editor.getModel(), selection = editor.getSelection(), selections = editor.getSelections() || [];
-                const position = editor.getPosition();
-                postMessageToFlutter({
-                  event: 'stats',
-                  lineCount: model.getLineCount(),
-                  charCount: model.getValueLength(),
-                  selLines: selection.isEmpty() ? 0 : (selection.endLineNumber - selection.startLineNumber + 1),
-                  selChars: selection.isEmpty() ? 0 : model.getValueInRange(selection).length,
-                  caretCount: selections.length,
-                  language: model.getLanguageId ? model.getLanguageId() : monaco.editor.getModelLanguage(model),
-                  cursorLine: position?.lineNumber,
-                  cursorColumn: position?.column,
-                });
-              };
-              editor.onDidChangeModelContent(sendStats);
-              editor.onDidChangeCursorSelection(sendStats);
-              sendStats();
-
-              // Set up typed API
-              (function () {
-                const E = () => window.editor;
-                const isMobileInputPlatform = () => {
-                  const ua = navigator.userAgent || '';
-                  return /Android|iPhone|iPad|iPod/i.test(ua) ||
-                    (navigator.platform === 'MacIntel' && navigator.maxTouchPoints > 1);
-                };
-                // Listeners this document registers on PARENT window objects
-                // (the Flutter host page) outlive it when the host removes
-                // the iframe element: removal discards the browsing context
-                // WITHOUT firing pagehide, so the parent-side EventTarget
-                // keeps the handler - and the handler roots this whole
-                // document plus its Monaco instance - indefinitely. Every
-                // parent-side registration therefore goes through this
-                // helper: the wrapper drops itself on the first event after
-                // the frame leaves the parent DOM, pagehide detaches
-                // everything on in-place navigation (the load-retry path),
-                // and the Dart controller calls
-                // __flutterMonacoDetachParentBindings from dispose() so
-                // cleanup does not wait for the next parent viewport event.
-                const parentBindings = [];
-                const bindParentListener = (target, type, handler, options) => {
-                  if (!target) return;
-                  const unbind = () => {
-                    try { target.removeEventListener(type, wrapped, options); } catch (_) {}
-                  };
-                  const wrapped = (event) => {
-                    const frame = window.frameElement;
-                    if (!frame || !frame.isConnected) {
-                      unbind();
-                      return;
-                    }
-                    handler(event);
-                  };
-                  try {
-                    target.addEventListener(type, wrapped, options);
-                    parentBindings.push(unbind);
-                  } catch (_) {}
-                };
-                window.__flutterMonacoDetachParentBindings = () => {
-                  while (parentBindings.length) {
-                    const unbind = parentBindings.pop();
-                    try { unbind(); } catch (_) {}
-                  }
-                };
-                window.addEventListener('pagehide', window.__flutterMonacoDetachParentBindings, { once: true });
-                const getEditorNode = () => {
-                  const ed = E();
-                  if (!ed) return null;
-                  return (ed.getDomNode && ed.getDomNode()) ||
-                    (ed.getContainerDomNode && ed.getContainerDomNode()) ||
-                    null;
-                };
-                const focusEditorTextAreaNow = () => {
-                  try {
-                    const ed = E();
-                    const node = getEditorNode();
-                    if (!ed || !node) return;
-                    try { ed.layout && ed.layout(); } catch (_) {}
-                    try { ed.focus && ed.focus(); } catch (_) {}
-                    try {
-                      const ta = node.querySelector('textarea.inputarea');
-                      if (ta && document.activeElement !== ta) {
-                        ta.focus();
-                      }
-                    } catch (_) {}
-                  } catch (_) {}
-                };
-                const serialize = (obj) => JSON.stringify(obj);
-                
-
-                // Events -> Flutter
-                const post = (event, payload) =>
-                  postMessageToFlutter({ event, ...payload });
-
-                E().onDidChangeModelContent(e => post('contentChanged', { isFlush: e.isFlush }));
-                E().onDidChangeCursorSelection(e => post('selectionChanged', {
-                  selection: e.selection && {
-                    startLineNumber: e.selection.startLineNumber,
-                    startColumn: e.selection.startColumn,
-                    endLineNumber: e.selection.endLineNumber,
-                    endColumn: e.selection.endColumn
-                  }
-                }));
-                E().onDidFocusEditorWidget(() => post('focus', {}));
-                E().onDidBlurEditorWidget(() => post('blur', {}));
-
-                // Typed helpers Flutter will call
-                const escapeRegExp = (value) =>
-                  (value ?? '').replace(/$jsEscapePattern/g, '\\\\\$&');
-
-                // Strict accessors used by the new flutterMonacoInvoke envelope.
-                // Helpers that depend on the editor/model should call these so
-                // missing-state errors propagate to Dart instead of being
-                // silently swallowed.
-                const requireEditor = () => {
-                  const ed = E();
-                  if (!ed) {
-                    throw new Error('Monaco editor is not ready.');
-                  }
-                  return ed;
-                };
-                const requireModel = () => {
-                  const ed = requireEditor();
-                  const model = ed.getModel ? ed.getModel() : null;
-                  if (!model) {
-                    throw new Error('Monaco editor has no active model.');
-                  }
-                  return model;
-                };
-
-                // Bridge dispatcher with a result envelope.
-                //
-                // Dart-side _invokeMonacoCommand calls this so that any
-                // JavaScript error inside a flutterMonaco helper is captured
-                // as a structured failure instead of crossing the WebView
-                // boundary as an uncaught exception.
-                //
-                // Success: { __flutterMonacoEval: true, ok: true, isUndefined, value }
-                // Failure: { __flutterMonacoEval: true, ok: false, error: { name, message, stack } }
-                window.flutterMonacoInvoke = (method, args) => {
-                  try {
-                    const api = window.flutterMonaco;
-                    const fn = api && api[method];
-                    if (typeof fn !== 'function') {
-                      throw new Error('Unknown flutterMonaco method: ' + method);
-                    }
-                    const value = fn.apply(api, Array.isArray(args) ? args : []);
-                    return {
-                      __flutterMonacoEval: true,
-                      ok: true,
-                      isUndefined: typeof value === 'undefined',
-                      value: typeof value === 'undefined' ? null : value,
-                    };
-                  } catch (e) {
-                    console.error('[flutterMonaco] invoke failed:', method, e);
-                    return {
-                      __flutterMonacoEval: true,
-                      ok: false,
-                      error: {
-                        name: e && e.name ? String(e.name) : 'Error',
-                        message: e && e.message ? String(e.message) : String(e),
-                        stack: e && e.stack ? String(e.stack) : null,
-                      },
-                    };
-                  }
-                };
-
-                // Async bridge dispatcher.
-                //
-                // flutterMonacoInvoke returns synchronously, which cannot
-                // represent promise-returning helpers (LSP connect must await
-                // the initialize handshake). This variant resolves the method
-                // (dotted paths reach sub-namespaces like 'lsp.connect'),
-                // awaits the result, and reports back over flutterChannel as
-                // an 'invokeResult' event that Dart correlates by requestId.
-                // This behaves identically on Android, iOS, macOS, Windows,
-                // and Web because no platform has to unwrap a JS Promise.
-                window.flutterMonacoInvokeAsync = (requestId, method, args) => {
-                  const respond = (payload) => {
-                    postMessageToFlutter(Object.assign(
-                      { event: 'invokeResult', requestId: requestId }, payload));
-                  };
-                  Promise.resolve()
-                    .then(() => {
-                      let parent = null;
-                      let target = window.flutterMonaco;
-                      for (const part of String(method).split('.')) {
-                        parent = target;
-                        target = target ? target[part] : undefined;
-                      }
-                      if (typeof target !== 'function') {
-                        throw new Error('Unknown flutterMonaco method: ' + method);
-                      }
-                      return target.apply(parent, Array.isArray(args) ? args : []);
-                    })
-                    .then((value) => respond({
-                      ok: true,
-                      isUndefined: typeof value === 'undefined',
-                      value: typeof value === 'undefined' ? null : value,
-                    }))
-                    .catch((e) => {
-                      console.error('[flutterMonaco] async invoke failed:', method, e);
-                      respond({
-                        ok: false,
-                        error: {
-                          name: e && e.name ? String(e.name) : 'Error',
-                          message: e && e.message ? String(e.message) : String(e),
-                          stack: e && e.stack ? String(e.stack) : null,
-                        },
-                      });
-                    });
-                  return true;
-                };
-
-                window.flutterMonaco = {
-
-                  // Basic editor operations
-                  focus: () => E().focus(),
-                  layout: () => { try { E().layout(); } catch (_) {} },
-                  // Force focus robustly: wait for visibility, layout, focus editor and hidden textarea
-                  forceFocus: (options = {}) => {
-                    try {
-                      const ed = E();
-                      const node = getEditorNode();
-                      if (!ed || !node) return;
-                      const replayInputFocus =
-                        options && options.replayInputFocus === true;
-
-                      if (isMobileInputPlatform()) {
-                        focusEditorTextAreaNow();
-                        return;
-                      }
-
-                      const attempt = () => {
-                        const rect = node.getBoundingClientRect();
-                        if (!rect.width || !rect.height) {
-                          return requestAnimationFrame(attempt);
-                        }
-                        // Idempotency guard: if the editor's input already owns
-                        // focus, return WITHOUT the document.body.focus() handoff
-                        // below. That handoff blurs then refocuses, which flickers
-                        // the caret and tears down an open context menu. Refocusing
-                        // an already-focused editor must be a no-op - this removes
-                        // the flicker for EVERY caller, not just guarded pointers.
-                        const input = node.querySelector(
-                          'textarea.inputarea, .native-edit-context');
-                        if (!replayInputFocus && input && document.activeElement === input) {
-                          return;
-                        }
-                        try { window.focus && window.focus(); } catch (_) {}
-                        try {
-                          if (document.body && !document.body.hasAttribute('tabindex')) {
-                            document.body.setAttribute('tabindex', '-1');
-                          }
-                          document.body?.focus?.();
-                        } catch (_) {}
-                        try { ed.layout && ed.layout(); } catch (_) {}
-                        try { ed.focus && ed.focus(); } catch (_) {}
-                        try {
-                          const ta = node.querySelector('textarea.inputarea');
-                          if (ta && document.activeElement !== ta) {
-                            ta.focus({ preventScroll: true });
-                            setTimeout(() => {
-                              try { ta.focus({ preventScroll: true }); } catch (_) {}
-                            }, 16);
-                          }
-                        } catch (_) {}
-                      };
-                      setTimeout(() => requestAnimationFrame(attempt), 0);
-                    } catch (_) {}
-                  },
-                  getValue: () => requireEditor().getValue(),
-                  setValue: (v) => {
-                    requireEditor().setValue(v || '');
-                    return true;
-                  },
-                  defineTheme: (name, data) => {
-                    if (!window.monaco || !monaco.editor) {
-                      throw new Error('Monaco editor API is not available.');
-                    }
-                    if (!name || typeof name !== 'string') {
-                      throw new Error('Theme name must be a non-empty string.');
-                    }
-                    monaco.editor.defineTheme(name, data || {});
-                    return true;
-                  },
-                  setHostPageBackground: (color) => {
-                    if (!color) {
-                      throw new Error('Host page background color is required.');
-                    }
-                    const value = String(color);
-                    document.documentElement.style.backgroundColor = value;
-                    document.body.style.backgroundColor = value;
-                    const container = document.getElementById('editor-container');
-                    if (container) container.style.backgroundColor = value;
-                    return true;
-                  },
-                  setTheme: (theme) => {
-                    if (!theme || typeof theme !== 'string') {
-                      throw new Error('Theme id must be a non-empty string.');
-                    }
-                    monaco.editor.setTheme(theme);
-                    return true;
-                  },
-                  getTheme: () => {
-                    if (!window.monaco || !monaco.editor ||
-                        typeof monaco.editor.getTheme !== 'function') {
-                      return null;
-                    }
-                    const theme = monaco.editor.getTheme();
-                    if (!theme) return null;
-                    if (typeof theme === 'string') return theme;
-                    return theme.themeName || theme.id || null;
-                  },
-                  setLanguage: (lang) => {
-                    monaco.editor.setModelLanguage(requireModel(), lang);
-                    return true;
-                  },
-                  updateOptions: (opts) => {
-                    requireEditor().updateOptions(opts);
-                    return true;
-                  },
-                  executeAction: (actionId, args) => {
-                    if (!actionId || typeof actionId !== 'string') {
-                      throw new Error('Action id must be a non-empty string.');
-                    }
-                    const ed = requireEditor();
-                    const action = ed.getAction ? ed.getAction(actionId) : null;
-                    if (action && typeof action.run === 'function') {
-                      action.run(args);
-                      return true;
-                    }
-                    ed.trigger('flutter-bridge', actionId, args);
-                    return true;
-                  },
-
-                  // Selection
-                  getSelection: () => {
-                    const s = E().getSelection();
-                    return s ? {
-                      startLineNumber: s.startLineNumber, startColumn: s.startColumn,
-                      endLineNumber: s.endLineNumber, endColumn: s.endColumn
-                    } : null;
-                  },
-                  setSelection: (r) => {
-                    requireEditor().setSelection(r);
-                    return true;
-                  },
-
-                  // Cursor
-                  getCursorPosition: () => {
-                    const p = E().getPosition();
-                    return p ? { lineNumber: p.lineNumber, column: p.column } : null;
-                  },
-                  setCursorPosition: (line, column) => {
-                    requireEditor().setPosition({ lineNumber: line, column: column });
-                    return true;
-                  },
-
-                  // Navigation
-                  revealLine: (ln, center) =>
-                    center ? E().revealLineInCenter(ln) : E().revealLine(ln),
-                  revealRange: (r, center) =>
-                    center ? E().revealRangeInCenter(r) : E().revealRange(r),
-
-                  // Line operations
-                  getLineCount: () => requireModel().getLineCount(),
-                  getLineContent: (ln) => requireModel().getLineContent(ln),
-                  
-                  // Word lookup
-                  getWordAtPosition: (line, column) => {
-                    const m = E().getModel();
-                    if (!m) return null;
-                    const w = m.getWordAtPosition(new monaco.Position(line, column));
-                    return w ? w.word : null;
-                  },
-
-                  // Edits
-                  applyEdits: (edits, opts) => {
-                    requireModel().applyEdits(edits || [], opts || {});
-                    return true;
-                  },
-
-                  // Decorations
-                  deltaDecorations: (oldIds, newDecos) =>
-                    requireEditor().deltaDecorations(oldIds || [], newDecos || []),
-
-                  // JSON language diagnostics.
-                  // Monaco 0.55 moved the language defaults from
-                  // monaco.languages.json to the top-level monaco.json and
-                  // deprecated the old path; prefer the new namespace and
-                  // fall back so a vendored older build keeps working.
-                  setJsonDiagnosticsOptions: (diagnostics) => {
-                    const jsonApi = (typeof monaco.json !== 'undefined' && monaco.json) ||
-                      (monaco.languages && monaco.languages.json);
-                    if (!jsonApi || !jsonApi.jsonDefaults) {
-                      throw new Error('Monaco JSON language API is not available.');
-                    }
-                    jsonApi.jsonDefaults.setDiagnosticsOptions(diagnostics);
-                    return true;
-                  },
-
-                  // Markers (diagnostics)
-                  setModelMarkers: (owner, markers) => {
-                    monaco.editor.setModelMarkers(requireModel(), owner || 'flutter', markers || []);
-                    return true;
-                  },
-
-                  // Find & replace (programmatic)
-                  findMatches: (q, options, limit) => {
-                    const m = E().getModel();
-                    if (!m) return [];
-                    const isRegex = !!(options && options.isRegex);
-                    const matchCase = !!(options && options.matchCase);
-                    const wholeWord = !!(options && options.wholeWord);
-
-                    let search = q ?? '';
-                    let useRegex = isRegex;
-                    if (wholeWord && !isRegex) {
-                      search = '\\\\b' + escapeRegExp(String(q ?? '')) + '\\\\b';
-                      useRegex = true;
-                    }
-
-                    const matches = m.findMatches(
-                      search,
-                      null,                 // searchScope: null = whole model (FIX: was 'false')
-                      useRegex,
-                      matchCase,
-                      null,
-                      false,                // captureMatches
-                      limit || 9999
-                    );
-                    return matches.map(mm => ({ range: mm.range, match: m.getValueInRange(mm.range) }));
-                  },
-
-                  replaceMatches: (q, repl, options) => {
-                    const m = E().getModel();
-                    if (!m) return 0;
-                    const isRegex = !!(options && options.isRegex);
-                    const matchCase = !!(options && options.matchCase);
-                    const wholeWord = !!(options && options.wholeWord);
-
-                    let search = q ?? '';
-                    let useRegex = isRegex;
-                    if (wholeWord && !isRegex) {
-                      search = '\\\\b' + escapeRegExp(String(q ?? '')) + '\\\\b';
-                      useRegex = true;
-                    }
-
-                    const matches = m.findMatches(
-                      search,
-                      null,                 // searchScope: null = whole model (FIX: was 'false')
-                      useRegex,
-                      matchCase,
-                      null,
-                      false,                // captureMatches
-                      9999
-                    );
-                    const edits = matches.map(mm => ({ range: mm.range, text: repl }));
-                    m.pushEditOperations([], edits, () => null);
-                    return edits.length;
-                  },
-
-                  // View state
-                  saveViewState: () => E().saveViewState(),
-                  restoreViewState: (s) => E().restoreViewState(s),
-
-                  // Stats
-                  getStatistics: () => {
-                    const e = E(), m = e.getModel(), s = e.getSelection();
-                    const position = e.getPosition(); // FIX: Define position in this scope
-                    const selections = e.getSelections() || [];
-                    return {
-                      lineCount: m ? m.getLineCount() : 0,
-                      charCount: m ? m.getValueLength() : 0,
-                      selLines: (s && !s.isEmpty()) ? (s.endLineNumber - s.startLineNumber + 1) : 0, // FIX: Return 0 for empty selection
-                      selChars: (m && s && !s.isEmpty()) ? m.getValueInRange(s).length : 0,
-                      caretCount: selections.length,
-                      language: m ? (m.getLanguageId ? m.getLanguageId() : monaco.editor.getModelLanguage(m)) : undefined,
-                      cursorLine: position?.lineNumber,
-                      cursorColumn: position?.column,
-                    };
-                  },
-                  
-                  // Dirty tracking (per-model baselines keyed by URI)
-                  _baselines: new Map(),
-                  _markBaseline(model) {
-                    const m = model || E().getModel();
-                    if (m && m.uri) {
-                      this._baselines.set(m.uri.toString(), m.getAlternativeVersionId());
-                    }
-                  },
-                  hasUnsavedChanges: () => {
-                    const m = E().getModel();
-                    if (!m || !m.uri) return false;
-                    const uri = m.uri.toString();
-                    if (!window.flutterMonaco._baselines.has(uri)) {
-                      window.flutterMonaco._markBaseline(m);
-                    }
-                    return m.getAlternativeVersionId() !== window.flutterMonaco._baselines.get(uri);
-                  },
-                  markSaved: () => window.flutterMonaco._markBaseline(),
-
-                  // Models
-                  createModel: (value, language, uri) =>
-                    monaco.editor.createModel(value || '', language || 'plaintext', 
-                      uri ? monaco.Uri.parse(uri) : undefined).uri.toString(),
-                  setModel: (uri) => {
-                    const m = monaco.editor.getModel(monaco.Uri.parse(uri));
-                    if (m) E().setModel(m);
-                  },
-                  disposeModel: (uri) => {
-                    const m = monaco.editor.getModel(monaco.Uri.parse(uri));
-                    if (m) m.dispose();
-                  },
-                  listModels: () => monaco.editor.getModels().map(m => m.uri.toString()),
-                };
-
-                if (isMobileInputPlatform()) {
-                  try {
-                    const node = getEditorNode();
-                    if (node && !node.__flutterMonacoMobileFocusBound) {
-                      node.__flutterMonacoMobileFocusBound = true;
-                      const ownerDocument = node.ownerDocument || document;
-                      const ownerWindow = ownerDocument.defaultView || window;
-                      const ua = navigator.userAgent || '';
-                      const isAndroid = /Android/i.test(ua);
-                      const isFlutterWebEmbed = (() => {
-                        try {
-                          return ownerWindow.parent && ownerWindow.parent !== ownerWindow;
-                        } catch (_) {
-                          return false;
-                        }
-                      })();
-                      const tapMoveThreshold = 8;
-                      const tapTimeThreshold = 650;
-                      const compatibilityEventSuppressMs = 1200;
-                      let gesture = null;
-                      let androidTouchScrollGesture = null;
-                      let lastFocusAt = 0;
-                      let suppressUntil = 0;
-                      let suppressFocusUntil = 0;
-                      const supportsPointerEvents = !!ownerWindow.PointerEvent;
-                      const usePointerTapBridge = supportsPointerEvents && isAndroid;
-                      const useTouchTapBridge = !usePointerTapBridge;
-                      const useAndroidWebFocusGuard = usePointerTapBridge && isFlutterWebEmbed;
-                      const now = () => Date.now();
-                      const eventPoint = (event) => {
-                        const touch =
-                          event.changedTouches?.[0] || event.touches?.[0];
-                        if (touch) {
-                          return { x: touch.clientX, y: touch.clientY };
-                        }
-                        if (typeof event.clientX === 'number') {
-                          return { x: event.clientX, y: event.clientY };
-                        }
-                        return null;
-                      };
-                      const getScrollSnapshot = () => {
-                        try {
-                          const ed = E();
-                          if (!ed || !ed.getScrollTop || !ed.getScrollLeft) {
-                            return { top: 0, left: 0 };
-                          }
-                          return {
-                            top: ed.getScrollTop(),
-                            left: ed.getScrollLeft(),
-                          };
-                        } catch (_) {
-                          return { top: 0, left: 0 };
-                        }
-                      };
-                      const hasMovedFromStart = (event) => {
-                        if (!gesture) return false;
-                        const point = eventPoint(event);
-                        if (point) {
-                          const dx = point.x - gesture.x;
-                          const dy = point.y - gesture.y;
-                          if ((dx * dx + dy * dy) > tapMoveThreshold * tapMoveThreshold) {
-                            return true;
-                          }
-                        }
-                        const scroll = getScrollSnapshot();
-                        return Math.abs(scroll.top - gesture.scrollTop) > 0 ||
-                          Math.abs(scroll.left - gesture.scrollLeft) > 0;
-                      };
-                      const hasTouchScrollMovedFromStart = (event) => {
-                        if (!androidTouchScrollGesture) return false;
-                        const point = eventPoint(event);
-                        if (point) {
-                          const dx = point.x - androidTouchScrollGesture.x;
-                          const dy = point.y - androidTouchScrollGesture.y;
-                          if ((dx * dx + dy * dy) > tapMoveThreshold * tapMoveThreshold) {
-                            return true;
-                          }
-                        }
-                        const scroll = getScrollSnapshot();
-                        return Math.abs(scroll.top - androidTouchScrollGesture.scrollTop) > 0 ||
-                          Math.abs(scroll.left - androidTouchScrollGesture.scrollLeft) > 0;
-                      };
-                      const blockEvent = (event) => {
-                        try {
-                          if (event.cancelable && event.preventDefault) {
-                            event.preventDefault();
-                          }
-                        } catch (_) {}
-                        try { event.stopImmediatePropagation && event.stopImmediatePropagation(); } catch (_) {}
-                        try { event.stopPropagation && event.stopPropagation(); } catch (_) {}
-                      };
-                      const suppressCompatibilityEvents = () => {
-                        suppressUntil = now() + compatibilityEventSuppressMs;
-                      };
-                      const suppressAndBlock = (event) => {
-                        suppressCompatibilityEvents();
-                        blockEvent(event);
-                      };
-                      const shouldBlockSuppressedEvent = () => now() < suppressUntil;
-                      const blockSuppressedCompatibilityEvent = (event) => {
-                        if (shouldBlockSuppressedEvent()) {
-                          blockEvent(event);
-                        }
-                      };
-                      const editorInputSelector =
-                        'textarea.inputarea, .native-edit-context';
-                      const isEditorInputElement = (element) => {
-                        try {
-                          return !!(
-                            element &&
-                            element.matches &&
-                            element.matches(editorInputSelector)
-                          );
-                        } catch (_) {
-                          return false;
-                        }
-                      };
-                      const getEditorInputElement = () => {
-                        try {
-                          const active = ownerDocument.activeElement;
-                          if (isEditorInputElement(active)) {
-                            return active;
-                          }
-                          return node.querySelector(editorInputSelector);
-                        } catch (_) {
-                          return null;
-                        }
-                      };
-                      const shouldSuppressFocus = () => now() < suppressFocusUntil;
-                      let maxObservedViewportHeight = 0;
-                      const getViewportHeightForKeyboard = () => {
-                        const readHeight = (win) => {
-                          try {
-                            return win?.visualViewport?.height || win?.innerHeight || 0;
-                          } catch (_) {
-                            return 0;
-                          }
-                        };
-                        let height = readHeight(ownerWindow);
-                        try {
-                          if (ownerWindow.parent && ownerWindow.parent !== ownerWindow) {
-                            height = readHeight(ownerWindow.parent) || height;
-                          }
-                        } catch (_) {}
-                        return height || 0;
-                      };
-                      const updateViewportKeyboardBaseline = () => {
-                        const height = getViewportHeightForKeyboard();
-                        if (height > maxObservedViewportHeight) {
-                          maxObservedViewportHeight = height;
-                        }
-                        return height;
-                      };
-                      const isKeyboardLikelyVisible = () => {
-                        const height = updateViewportKeyboardBaseline();
-                        const baseline = maxObservedViewportHeight || height;
-                        if (!height || !baseline) return false;
-                        const hiddenHeight = baseline - height;
-                        return hiddenHeight > Math.max(120, baseline * 0.18);
-                      };
-                      updateViewportKeyboardBaseline();
-                      try {
-                        ownerWindow.visualViewport?.addEventListener(
-                          'resize',
-                          updateViewportKeyboardBaseline,
-                          { passive: true }
-                        );
-                      } catch (_) {}
-                      try {
-                        if (ownerWindow.parent && ownerWindow.parent !== ownerWindow) {
-                          bindParentListener(
-                            ownerWindow.parent.visualViewport,
-                            'resize',
-                            updateViewportKeyboardBaseline,
-                            { passive: true }
-                          );
-                        }
-                      } catch (_) {}
-                      const blurTextAreaIfFocusSuppressed = () => {
-                        if (!shouldSuppressFocus()) return;
-                        try {
-                          const input = getEditorInputElement();
-                          if (input && ownerDocument.activeElement === input) {
-                            input.blur();
-                          }
-                        } catch (_) {}
-                      };
-                      const scheduleSuppressedTextAreaBlur = () => {
-                        blurTextAreaIfFocusSuppressed();
-                        try { ownerWindow.setTimeout(blurTextAreaIfFocusSuppressed, 0); } catch (_) {}
-                        try { ownerWindow.setTimeout(blurTextAreaIfFocusSuppressed, 50); } catch (_) {}
-                      };
-                      const suppressScrollFocusIfNeeded = () => {
-                        if (!useAndroidWebFocusGuard) return;
-                        suppressFocusUntil = now() + compatibilityEventSuppressMs;
-                        const keyboardVisible = isKeyboardLikelyVisible();
-                        if (!keyboardVisible) {
-                          scheduleSuppressedTextAreaBlur();
-                        }
-                      };
-                      const guardSuppressedTextAreaFocus = (event) => {
-                        if (!useAndroidWebFocusGuard || !shouldSuppressFocus()) return;
-                        const target = event?.target;
-                        if (isEditorInputElement(target)) {
-                          blurTextAreaIfFocusSuppressed();
-                          blockEvent(event);
-                        }
-                      };
-                      const beginGesture = (event, id, kind) => {
-                        const point = eventPoint(event);
-                        if (!point) return;
-                        const scroll = getScrollSnapshot();
-                        gesture = {
-                          id,
-                          kind,
-                          x: point.x,
-                          y: point.y,
-                          startedAt: now(),
-                          moved: false,
-                          cancelled: false,
-                          scrollTop: scroll.top,
-                          scrollLeft: scroll.left,
-                        };
-                      };
-                      const updateGesture = (event, id, kind) => {
-                        if (!gesture || gesture.id !== id || gesture.kind !== kind) return;
-                        if (hasMovedFromStart(event)) {
-                          gesture.moved = true;
-                          if (useAndroidWebFocusGuard) {
-                            suppressCompatibilityEvents();
-                            suppressScrollFocusIfNeeded();
-                          }
-                        }
-                      };
-                      const cancelGesture = (event, id, kind) => {
-                        if (!gesture || gesture.id !== id || gesture.kind !== kind) return;
-                        gesture.cancelled = true;
-                        suppressScrollFocusIfNeeded();
-                        suppressAndBlock(event);
-                        gesture = null;
-                      };
-                      const endGesture = (event, id, kind) => {
-                        if (!gesture || gesture.id !== id || gesture.kind !== kind) {
-                          if (shouldBlockSuppressedEvent()) {
-                            blockEvent(event);
-                          }
-                          return;
-                        }
-                        updateGesture(event, id, kind);
-                        const elapsed = now() - gesture.startedAt;
-                        const shouldFocus =
-                          !gesture.cancelled &&
-                          !gesture.moved &&
-                          elapsed <= tapTimeThreshold;
-                        gesture = null;
-                        if (!shouldFocus) {
-                          suppressScrollFocusIfNeeded();
-                          suppressAndBlock(event);
-                          return;
-                        }
-                        suppressUntil = 0;
-                        suppressFocusUntil = 0;
-                        const currentTime = now();
-                        if (currentTime - lastFocusAt < 150) return;
-                        lastFocusAt = currentTime;
-                        focusEditorTextAreaNow();
-                      };
-                      const pointerId = (event) =>
-                        typeof event.pointerId === 'number' ? event.pointerId : 1;
-                      const onPointerDown = (event) => {
-                        if (!event.isPrimary || event.pointerType !== 'touch') return;
-                        beginGesture(event, pointerId(event), 'pointer');
-                      };
-                      const onPointerMove = (event) => {
-                        if (!event.isPrimary || event.pointerType !== 'touch') return;
-                        updateGesture(event, pointerId(event), 'pointer');
-                      };
-                      const onPointerUp = (event) => {
-                        if (!event.isPrimary || event.pointerType !== 'touch') return;
-                        endGesture(event, pointerId(event), 'pointer');
-                      };
-                      const onPointerCancel = (event) => {
-                        if (!event.isPrimary || event.pointerType !== 'touch') return;
-                        cancelGesture(event, pointerId(event), 'pointer');
-                      };
-                      const firstChangedTouch = (event) => event.changedTouches?.[0] || null;
-                      const firstActiveTouch = (event) => event.touches?.[0] || null;
-                      const onTouchStart = (event) => {
-                        const touch = firstActiveTouch(event) || firstChangedTouch(event);
-                        if (!touch) return;
-                        beginGesture(event, touch.identifier, 'touch');
-                      };
-                      const onTouchMove = (event) => {
-                        const touch = firstChangedTouch(event) || firstActiveTouch(event);
-                        if (!touch) return;
-                        updateGesture(event, touch.identifier, 'touch');
-                      };
-                      const onTouchEnd = (event) => {
-                        const touch = firstChangedTouch(event);
-                        if (!touch) return;
-                        endGesture(event, touch.identifier, 'touch');
-                      };
-                      const onTouchCancel = (event) => {
-                        const touch = firstChangedTouch(event);
-                        if (!touch) return;
-                        cancelGesture(event, touch.identifier, 'touch');
-                      };
-                      const beginAndroidTouchScrollGuard = (event) => {
-                        if (!useAndroidWebFocusGuard) return;
-                        const touch = firstActiveTouch(event) || firstChangedTouch(event);
-                        if (!touch) return;
-                        const point = eventPoint(event);
-                        if (!point) return;
-                        const scroll = getScrollSnapshot();
-                        androidTouchScrollGesture = {
-                          id: touch.identifier,
-                          x: point.x,
-                          y: point.y,
-                          moved: false,
-                          scrollTop: scroll.top,
-                          scrollLeft: scroll.left,
-                        };
-                      };
-                      const updateAndroidTouchScrollGuard = (event) => {
-                        if (!useAndroidWebFocusGuard || !androidTouchScrollGesture) return;
-                        if (hasTouchScrollMovedFromStart(event)) {
-                          androidTouchScrollGesture.moved = true;
-                          suppressCompatibilityEvents();
-                          suppressScrollFocusIfNeeded();
-                        }
-                      };
-                      const endAndroidTouchScrollGuard = (event) => {
-                        if (!useAndroidWebFocusGuard) return;
-                        if (!androidTouchScrollGesture) {
-                          if (shouldBlockSuppressedEvent()) {
-                            blockEvent(event);
-                          }
-                          return;
-                        }
-                        updateAndroidTouchScrollGuard(event);
-                        const shouldBlock =
-                          androidTouchScrollGesture.moved ||
-                          shouldBlockSuppressedEvent();
-                        androidTouchScrollGesture = null;
-                        if (shouldBlock) {
-                          blockEvent(event);
-                        }
-                      };
-                      const cancelAndroidTouchScrollGuard = (event) => {
-                        if (!useAndroidWebFocusGuard || !androidTouchScrollGesture) return;
-                        suppressCompatibilityEvents();
-                        suppressScrollFocusIfNeeded();
-                        androidTouchScrollGesture = null;
-                        blockEvent(event);
-                      };
-                      const capturePassiveFalse = { capture: true, passive: false };
-                      const capturePassiveTrue = { capture: true, passive: true };
-                      const captureOnly = { capture: true };
-
-                      if (usePointerTapBridge) {
-                        ownerDocument.addEventListener('pointerdown', onPointerDown, captureOnly);
-                        ownerDocument.addEventListener('pointermove', onPointerMove, capturePassiveFalse);
-                        ownerDocument.addEventListener('pointerup', onPointerUp, capturePassiveFalse);
-                        ownerDocument.addEventListener('pointercancel', onPointerCancel, capturePassiveFalse);
-                      }
-
-                      if (useAndroidWebFocusGuard) {
-                        ownerDocument.addEventListener('touchstart', beginAndroidTouchScrollGuard, capturePassiveTrue);
-                        ownerDocument.addEventListener('touchmove', updateAndroidTouchScrollGuard, capturePassiveFalse);
-                        ownerDocument.addEventListener('touchend', endAndroidTouchScrollGuard, capturePassiveFalse);
-                        ownerDocument.addEventListener('touchcancel', cancelAndroidTouchScrollGuard, capturePassiveFalse);
-                        ownerDocument.addEventListener('focus', guardSuppressedTextAreaFocus, captureOnly);
-                        ownerDocument.addEventListener('focusin', guardSuppressedTextAreaFocus, captureOnly);
-                      }
-
-                      if (useTouchTapBridge) {
-                        ownerDocument.addEventListener('touchstart', onTouchStart, capturePassiveTrue);
-                        ownerDocument.addEventListener('touchmove', onTouchMove, capturePassiveFalse);
-                        ownerDocument.addEventListener('touchend', onTouchEnd, capturePassiveFalse);
-                        ownerDocument.addEventListener('touchcancel', onTouchCancel, capturePassiveFalse);
-                      }
-
-                      ownerDocument.addEventListener('mousedown', blockSuppressedCompatibilityEvent, captureOnly);
-                      ownerDocument.addEventListener('mouseup', blockSuppressedCompatibilityEvent, captureOnly);
-                      ownerDocument.addEventListener('click', blockSuppressedCompatibilityEvent, captureOnly);
-                      try {
-                        node.style.touchAction = node.style.touchAction || 'manipulation';
-                      } catch (_) {}
-                    }
-                  } catch (_) {}
-                }
-                ${isWeb ? '''
-
-                // Mobile web keyboard fit (issue #11): the soft keyboard
-                // shrinks only the host page's *visual* viewport. This iframe
-                // keeps its layout size, so Monaco keeps rendering lines under
-                // the keyboard while Safari pans the host page to chase the
-                // hidden caret textarea - the editor's top band leaves the
-                // screen and cannot be scrolled back while the keyboard is up.
-                // While the parent's visual viewport is constrained, pin
-                // #editor-container to the visible intersection of this frame
-                // (automaticLayout relayouts Monaco via its ResizeObserver);
-                // restore the stylesheet layout once it is unconstrained.
-                if (isMobileInputPlatform() && !window.__flutterMonacoViewportFitBound) {
-                  try {
-                    const fitWindow = window;
-                    const fitContainer = document.getElementById('editor-container');
-                    const fitFrame = fitWindow.frameElement;
-                    const fitParent = fitWindow.parent;
-                    if (fitContainer && fitFrame && fitParent && fitParent !== fitWindow && fitParent.visualViewport) {
-                      fitWindow.__flutterMonacoViewportFitBound = true;
-                      const fitViewport = fitParent.visualViewport;
-                      // The pin below exists FOR the constrained-visual-viewport
-                      // states only: the soft keyboard shrinking the visual
-                      // viewport, or Safari panning it (nonzero offsets) to
-                      // chase the caret. A frame that merely extends past the
-                      // layout viewport - an editor half-scrolled off inside a
-                      // scrollable Flutter page - must keep its stylesheet
-                      // layout, or its content would anchor to the screen band
-                      // instead of scrolling away with the page. Pinch zoom
-                      // also shrinks the visual viewport but pans freely;
-                      // pinning would fight the user's zoom, so scale != 1 is
-                      // excluded.
-                      const isViewportConstrained = () => {
-                        const scale = fitViewport.scale || 1;
-                        if (Math.abs(scale - 1) > 0.02) return false;
-                        if (fitViewport.offsetLeft > 1 || fitViewport.offsetTop > 1) return true;
-                        const layoutWidth = fitParent.innerWidth || 0;
-                        const layoutHeight = fitParent.innerHeight || 0;
-                        // 24px absorbs scrollbar and rounding noise; soft
-                        // keyboards and hardware-keyboard accessory bars are
-                        // all taller than that.
-                        return (
-                          (layoutHeight > 0 && layoutHeight - fitViewport.height > 24) ||
-                          (layoutWidth > 0 && layoutWidth - fitViewport.width > 24)
-                        );
-                      };
-                      let fitRaf = 0;
-                      let fitApplied = false;
-                      let fitLast = '';
-                      const clearViewportFit = () => {
-                        fitLast = '';
-                        if (!fitApplied) return;
-                        fitApplied = false;
-                        fitContainer.style.position = '';
-                        fitContainer.style.top = '';
-                        fitContainer.style.left = '';
-                        fitContainer.style.width = '';
-                        fitContainer.style.height = '';
-                      };
-                      const scheduleViewportFit = () => {
-                        if (fitRaf) return;
-                        try {
-                          fitRaf = fitWindow.requestAnimationFrame(applyViewportFit);
-                        } catch (_) {
-                          fitRaf = 0;
-                        }
-                      };
-                      const applyViewportFit = () => {
-                        fitRaf = 0;
-                        let rect = null;
-                        try { rect = fitFrame.getBoundingClientRect(); } catch (_) {}
-                        if (!rect || rect.width <= 0 || rect.height <= 0) {
-                          clearViewportFit();
-                          return;
-                        }
-                        if (!isViewportConstrained()) {
-                          clearViewportFit();
-                          return;
-                        }
-                        // Keep tracking while constrained: Flutter can move the
-                        // frame (scrolling, route animations) without firing
-                        // any parent viewport event.
-                        scheduleViewportFit();
-                        const viewLeft = fitViewport.offsetLeft;
-                        const viewTop = fitViewport.offsetTop;
-                        const viewWidth = fitViewport.width || fitParent.innerWidth || rect.width;
-                        const viewHeight = fitViewport.height || fitParent.innerHeight || rect.height;
-                        const left = Math.max(rect.left, viewLeft);
-                        const top = Math.max(rect.top, viewTop);
-                        const width = Math.min(rect.right, viewLeft + viewWidth) - left;
-                        const height = Math.min(rect.bottom, viewTop + viewHeight) - top;
-                        if (width >= rect.width - 1 && height >= rect.height - 1) {
-                          // Fully visible above the keyboard: the stylesheet
-                          // layout is already correct; the reschedule above
-                          // keeps watching while the viewport stays
-                          // constrained.
-                          clearViewportFit();
-                          return;
-                        }
-                        // Sub-48px intersections are mid-transition slivers;
-                        // keep the previous geometry instead of collapsing.
-                        if (width < 48 || height < 48) return;
-                        const next =
-                          Math.round(left - rect.left) + ':' + Math.round(top - rect.top) +
-                          ':' + Math.round(width) + ':' + Math.round(height);
-                        if (next === fitLast) return;
-                        fitLast = next;
-                        fitApplied = true;
-                        fitContainer.style.position = 'absolute';
-                        fitContainer.style.left = Math.round(left - rect.left) + 'px';
-                        fitContainer.style.top = Math.round(top - rect.top) + 'px';
-                        fitContainer.style.width = Math.round(width) + 'px';
-                        fitContainer.style.height = Math.round(height) + 'px';
-                        try {
-                          const ed = E();
-                          if (ed && ed.hasTextFocus && ed.hasTextFocus()) {
-                            const pos = ed.getPosition && ed.getPosition();
-                            if (pos && ed.revealPosition) ed.revealPosition(pos);
-                          }
-                        } catch (_) {}
-                      };
-                      bindParentListener(fitViewport, 'resize', scheduleViewportFit, { passive: true });
-                      bindParentListener(fitViewport, 'scroll', scheduleViewportFit, { passive: true });
-                      bindParentListener(fitParent, 'resize', scheduleViewportFit, { passive: true });
-                      scheduleViewportFit();
-                    }
-                  } catch (_) {}
-                }
-''' : ''}
-                // Completion bridge: JS stays dumb, Flutter drives the data
-                (function () {
-                  const completion = {
-                    resolvers: Object.create(null),
-                    providers: Object.create(null),
-                    nextId: 1,
-                  };
-
-                  function toIRange(r) {
-                    if (!r) return undefined;
-                    const sL = r.startLineNumber ?? r.startLine ?? r.from_line ?? r.start ?? 1;
-                    const sC = r.startColumn ?? r.startCol ?? r.sc ?? 1;
-                    const eL = r.endLineNumber ?? r.endLine ?? r.to_line ?? r.end ?? sL;
-                    const eC = r.endColumn ?? r.endCol ?? r.ec ?? sC;
-                    return {
-                      startLineNumber: sL,
-                      startColumn: sC,
-                      endLineNumber: eL,
-                      endColumn: eC,
-                    };
-                  }
-
-                  // cfg: { id?: string, languages: string[]|string, triggerCharacters?: string[] }
-                  window.flutterMonaco.registerCompletionSource = function (cfg) {
-                    const id = cfg?.id || 'flutter_' + completion.nextId++;
-                    const langs = Array.isArray(cfg?.languages)
-                      ? cfg.languages
-                      : [cfg?.languages ?? 'plaintext'];
-                    const triggerCharacters = cfg?.triggerCharacters || [];
-
-                    const provider = {
-                      triggerCharacters,
-                      provideCompletionItems: (model, position, context, token) =>
-                        new Promise((resolve) => {
-                          const reqId =
-                            id + ':' + Date.now() + ':' + Math.random().toString(36).slice(2);
-                          const lang =
-                            (model.getLanguageId && model.getLanguageId()) ||
-                            monaco.editor.getModelLanguage(model);
-                          const word = model.getWordUntilPosition(position);
-                          const defaultRange = {
-                            startLineNumber: position.lineNumber,
-                            startColumn: word.startColumn,
-                            endLineNumber: position.lineNumber,
-                            endColumn: word.endColumn,
-                          };
-                          completion.resolvers[reqId] = {
-                            resolve,
-                            defaultRange,
-                          };
-
-                          const payload = {
-                            event: 'completionRequest',
-                            providerId: id,
-                            requestId: reqId,
-                            language: lang,
-                            uri: model.uri?.toString(),
-                            position: {
-                              lineNumber: position.lineNumber,
-                              column: position.column,
-                            },
-                            defaultRange,
-                            lineText: model.getLineContent(position.lineNumber),
-                            triggerKind: context?.triggerKind ?? null,
-                            triggerCharacter: context?.triggerCharacter ?? null,
-                          };
-                          postMessageToFlutter(payload);
-
-                          token?.onCancellationRequested?.(() => {
-                            delete completion.resolvers[reqId];
-                            try {
-                              resolve({ suggestions: [] });
-                            } catch (_) {}
-                          });
-                        }),
-                    };
-
-                    const disposables = langs.map((l) =>
-                      monaco.languages.registerCompletionItemProvider(l, provider),
-                    );
-                    completion.providers[id] = { disposables };
-                    return id;
-                  };
-
-                  window.flutterMonaco.unregisterCompletionSource = function (id) {
-                    const p = completion.providers[id];
-                    if (p?.disposables) {
-                      for (const d of p.disposables) {
-                        try {
-                          d.dispose();
-                        } catch (_) {}
-                      }
-                    }
-                    delete completion.providers[id];
-                  };
-
-                  // Flutter -> JS: deliver completion results
-                  window.flutterMonaco.complete = function (requestId, payload) {
-                    const resolver = completion.resolvers[requestId];
-                    if (!resolver) return;
-                    const resolve = resolver.resolve;
-                    const fallbackRange = resolver.defaultRange;
-                    try {
-                      const items = (payload && payload.suggestions) || [];
-                      const defaultRange = payload?.defaultRange || fallbackRange;
-                      const mapped = items.map((it) => {
-                        let kind = it.kind;
-                        if (typeof kind === 'string') {
-                          kind = monaco.languages.CompletionItemKind[kind] ??
-                            monaco.languages.CompletionItemKind.Text;
-                        }
-                        let insertTextRules = it.insertTextRules;
-                        if (Array.isArray(insertTextRules)) {
-                          insertTextRules = insertTextRules.reduce((mask, rule) => {
-                            const value = monaco.languages.CompletionItemInsertTextRule[rule];
-                            return typeof value === 'number' ? (mask | value) : mask;
-                          }, 0);
-                        }
-                        return {
-                          label: it.label,
-                          insertText: it.insertText || it.label,
-                          kind: kind || monaco.languages.CompletionItemKind.Text,
-                          detail: it.detail,
-                          documentation: it.documentation,
-                          sortText: it.sortText,
-                          filterText: it.filterText,
-                          commitCharacters: it.commitCharacters,
-                          insertTextRules: insertTextRules || undefined,
-                          range: toIRange(it.range) || toIRange(defaultRange),
-                        };
-                      });
-                      resolve({
-                        suggestions: mapped,
-                        incomplete: !!payload?.isIncomplete,
-                      });
-                    } finally {
-                      delete completion.resolvers[requestId];
-                    }
-                  };
-                })();
-
-                // Edge scroll handoff (opt-in): forward scroll deltas the
-                // editor cannot consume to Flutter, which applies them to a
-                // host scrollable. Sources are toggled from Dart through
-                // flutterMonaco.setScrollHandoff; while disabled this module
-                // installs no DOM listeners and adds no per-event work, so
-                // the default editor behavior is untouched.
-                (function () {
-                  const EDGE_EPSILON = 1;
-                  const TOUCH_SLOP = 8;
-                  const handoffState = {
-                    wheel: false,
-                    touch: false,
-                    touchGesture: null,
-                    touchSelectionWatch: null,
-                  };
-
-                  const scrollMetrics = (ed) => {
-                    let layout = null;
-                    try { layout = ed.getLayoutInfo ? ed.getLayoutInfo() : null; } catch (_) {}
-                    const viewportHeight = (layout && layout.height) || 0;
-                    const viewportWidth = (layout && layout.width) || 0;
-                    const scrollTop = ed.getScrollTop ? ed.getScrollTop() : 0;
-                    const scrollLeft = ed.getScrollLeft ? ed.getScrollLeft() : 0;
-                    const scrollHeight = ed.getScrollHeight ? ed.getScrollHeight() : 0;
-                    const scrollWidth = ed.getScrollWidth ? ed.getScrollWidth() : 0;
-                    const maxTop = Math.max(0, scrollHeight - viewportHeight);
-                    const maxLeft = Math.max(0, scrollWidth - viewportWidth);
-                    return {
-                      maxTop: maxTop,
-                      viewportHeight: viewportHeight,
-                      atTop: scrollTop <= EDGE_EPSILON,
-                      atBottom: scrollTop >= maxTop - EDGE_EPSILON,
-                      atLeft: scrollLeft <= EDGE_EPSILON,
-                      atRight: scrollLeft >= maxLeft - EDGE_EPSILON,
-                    };
-                  };
-
-                  const editorLineHeight = (ed) => {
-                    try {
-                      const value = ed.getOption(monaco.editor.EditorOption.lineHeight);
-                      if (typeof value === 'number' && value > 0) return value;
-                    } catch (_) {}
-                    return 20;
-                  };
-
-                  const editorWheelDisabled = (ed) => {
-                    try {
-                      const scrollbar = ed.getOption(monaco.editor.EditorOption.scrollbar);
-                      return !!scrollbar && scrollbar.handleMouseWheel === false;
-                    } catch (_) {}
-                    return false;
-                  };
-
-                  // Default deny: only the main editor scroll region hands
-                  // off. Monaco-owned overlays (suggest list, hover, find,
-                  // menus, peek editors) keep their own wheel handling, and
-                  // widgets rendered outside the editor DOM never match.
-                  const HANDOFF_BLOCKED_WIDGETS = [
-                    '.suggest-widget', '.suggest-details-container',
-                    '.monaco-hover', '.parameter-hints-widget',
-                    '.find-widget', '.context-view', '.monaco-menu',
-                    '.quick-input-widget', '.zone-widget',
-                    '.peekview-widget', '.rename-box', '.monaco-inputbox',
-                  ].join(', ');
-                  const isMainScrollRegion = (ed, target) => {
-                    const editorDom = ed.getDomNode ? ed.getDomNode() : null;
-                    if (!editorDom) return false;
-                    let el = target;
-                    if (el && el.nodeType !== 1) el = el.parentElement;
-                    if (!el || !el.closest || !editorDom.contains(el)) return false;
-                    if (el.closest(HANDOFF_BLOCKED_WIDGETS)) return false;
-                    if (el.closest('.monaco-editor') !== editorDom) return false;
-                    const scrollable = el.closest('.monaco-scrollable-element');
-                    if (scrollable && !scrollable.classList.contains('editor-scrollable')) {
-                      return false;
-                    }
-                    return true;
-                  };
-
-                  const postHandoff = (source, deltaX, deltaY, metrics) => {
-                    post('scrollHandoff', {
-                      source: source,
-                      deltaX: deltaX,
-                      deltaY: deltaY,
-                      atTop: metrics.atTop,
-                      atBottom: metrics.atBottom,
-                      atLeft: metrics.atLeft,
-                      atRight: metrics.atRight,
-                    });
-                  };
-
-                  const onHandoffWheel = (event) => {
-                    try {
-                      if (event.defaultPrevented) return;
-                      // Zoom owns ctrl/meta wheel: editor mouseWheelZoom,
-                      // browser zoom, and macOS pinch (reported as ctrl).
-                      if (event.ctrlKey || event.metaKey) return;
-                      const ed = E();
-                      if (!ed || !isMainScrollRegion(ed, event.target)) return;
-
-                      const metrics = scrollMetrics(ed);
-                      let deltaX = event.deltaX || 0;
-                      let deltaY = event.deltaY || 0;
-                      if (event.deltaMode === 1) {
-                        const lineHeight = editorLineHeight(ed);
-                        deltaX *= lineHeight;
-                        deltaY *= lineHeight;
-                      } else if (event.deltaMode === 2) {
-                        deltaX *= metrics.viewportHeight;
-                        deltaY *= metrics.viewportHeight;
-                      }
-                      // Vertical handoff only: dominant-horizontal wheel
-                      // input stays with the editor.
-                      if (deltaY === 0 || Math.abs(deltaY) <= Math.abs(deltaX)) return;
-
-                      const wantsDown = deltaY > 0;
-                      const editorCannotConsume =
-                        editorWheelDisabled(ed) ||
-                        metrics.maxTop <= EDGE_EPSILON ||
-                        (wantsDown ? metrics.atBottom : metrics.atTop);
-                      if (!editorCannotConsume) return;
-
-                      event.preventDefault();
-                      event.stopPropagation();
-                      postHandoff('wheel', deltaX, deltaY, metrics);
-                    } catch (_) {}
-                  };
-
-                  // Touch forwarding is observation-only: passive listeners
-                  // that never block, refocus, or select, so Monaco's touch
-                  // handling and the mobile focus guards keep working exactly
-                  // as before. Worst case is an extra parent scroll, never a
-                  // broken editor gesture.
-                  const gestureTouch = (event) => {
-                    const gesture = handoffState.touchGesture;
-                    if (!gesture) return null;
-                    const touches = event.changedTouches || [];
-                    for (let i = 0; i < touches.length; i++) {
-                      if (touches[i].identifier === gesture.id) return touches[i];
-                    }
-                    return null;
-                  };
-                  const onHandoffTouchStart = (event) => {
-                    try {
-                      if (event.touches && event.touches.length > 1) {
-                        handoffState.touchGesture = null;
-                        return;
-                      }
-                      const touchPoint = (event.touches && event.touches[0]) ||
-                        (event.changedTouches && event.changedTouches[0]);
-                      if (!touchPoint) return;
-                      const ed = E();
-                      if (!ed || !isMainScrollRegion(ed, event.target)) {
-                        handoffState.touchGesture = null;
-                        return;
-                      }
-                      handoffState.touchGesture = {
-                        id: touchPoint.identifier,
-                        startX: touchPoint.clientX,
-                        startY: touchPoint.clientY,
-                        lastY: touchPoint.clientY,
-                        moved: false,
-                        cancelled: false,
-                      };
-                    } catch (_) {}
-                  };
-                  const onHandoffTouchMove = (event) => {
-                    try {
-                      const gesture = handoffState.touchGesture;
-                      const touchPoint = gestureTouch(event);
-                      if (!gesture || !touchPoint || gesture.cancelled) return;
-                      const dy = touchPoint.clientY - gesture.lastY;
-                      gesture.lastY = touchPoint.clientY;
-                      if (!gesture.moved) {
-                        const totalX = touchPoint.clientX - gesture.startX;
-                        const totalY = touchPoint.clientY - gesture.startY;
-                        if ((totalX * totalX + totalY * totalY) <= TOUCH_SLOP * TOUCH_SLOP) {
-                          return;
-                        }
-                        gesture.moved = true;
-                      }
-                      if (dy === 0) return;
-                      const ed = E();
-                      if (!ed) return;
-                      // A downward finger drag requests scrolling up; express
-                      // the intent with wheel-style sign conventions.
-                      const deltaY = -dy;
-                      const metrics = scrollMetrics(ed);
-                      const wantsDown = deltaY > 0;
-                      const editorCannotConsume =
-                        metrics.maxTop <= EDGE_EPSILON ||
-                        (wantsDown ? metrics.atBottom : metrics.atTop);
-                      if (!editorCannotConsume) return;
-                      postHandoff('touch', 0, deltaY, metrics);
-                    } catch (_) {}
-                  };
-                  const onHandoffTouchEnd = (event) => {
-                    try {
-                      if (gestureTouch(event)) handoffState.touchGesture = null;
-                    } catch (_) {}
-                  };
-
-                  const wheelOptions = { capture: true, passive: false };
-                  const touchOptions = { capture: true, passive: true };
-                  const installScrollHandoffWheel = () => {
-                    window.addEventListener('wheel', onHandoffWheel, wheelOptions);
-                  };
-                  const removeScrollHandoffWheel = () => {
-                    window.removeEventListener('wheel', onHandoffWheel, wheelOptions);
-                  };
-                  const installScrollHandoffTouch = () => {
-                    window.addEventListener('touchstart', onHandoffTouchStart, touchOptions);
-                    window.addEventListener('touchmove', onHandoffTouchMove, touchOptions);
-                    window.addEventListener('touchend', onHandoffTouchEnd, touchOptions);
-                    window.addEventListener('touchcancel', onHandoffTouchEnd, touchOptions);
-                    try {
-                      const ed = E();
-                      if (ed && ed.onDidChangeCursorSelection && !handoffState.touchSelectionWatch) {
-                        // Text selection wins: a selection change during a
-                        // moved gesture stops forwarding for that gesture.
-                        handoffState.touchSelectionWatch = ed.onDidChangeCursorSelection(() => {
-                          const gesture = handoffState.touchGesture;
-                          if (gesture && gesture.moved) gesture.cancelled = true;
-                        });
-                      }
-                    } catch (_) {}
-                  };
-                  const removeScrollHandoffTouch = () => {
-                    window.removeEventListener('touchstart', onHandoffTouchStart, touchOptions);
-                    window.removeEventListener('touchmove', onHandoffTouchMove, touchOptions);
-                    window.removeEventListener('touchend', onHandoffTouchEnd, touchOptions);
-                    window.removeEventListener('touchcancel', onHandoffTouchEnd, touchOptions);
-                    const watch = handoffState.touchSelectionWatch;
-                    handoffState.touchSelectionWatch = null;
-                    if (watch && watch.dispose) {
-                      try { watch.dispose(); } catch (_) {}
-                    }
-                    handoffState.touchGesture = null;
-                  };
-
-                  window.flutterMonaco.setScrollHandoff = function (cfg) {
-                    const wheel = !!(cfg && cfg.wheel);
-                    const touch = !!(cfg && cfg.touch);
-                    if (wheel !== handoffState.wheel) {
-                      handoffState.wheel = wheel;
-                      if (wheel) { installScrollHandoffWheel(); }
-                      else { removeScrollHandoffWheel(); }
-                    }
-                    if (touch !== handoffState.touch) {
-                      handoffState.touch = touch;
-                      if (touch) { installScrollHandoffTouch(); }
-                      else { removeScrollHandoffTouch(); }
-                    }
-                    return true;
-                  };
-                })();
-
-                // LSP bridge: Dart owns transports and connection lifecycle,
-                // Monaco's built-in monaco.lsp.MonacoLspClient owns every
-                // language feature (completions, hover, diagnostics, rename,
-                // formatting, semantic tokens, ...). Nothing here mirrors LSP
-                // capabilities into Dart - once a connection is open the
-                // editor "just works".
-                (function () {
-                  const connections = new Map(); // id -> entry
-
-                  const lspAvailable = () =>
-                    !!(window.monaco && monaco.lsp && monaco.lsp.MonacoLspClient);
-
-                  function toErrorInfo(error) {
-                    if (!error) return null;
-                    return {
-                      name: error.name ? String(error.name) : 'Error',
-                      message: error.message ? String(error.message) : String(error),
-                    };
-                  }
-
-                  function postStatus(id, status, error) {
-                    postMessageToFlutter({
-                      event: 'lspStatus',
-                      connectionId: id,
-                      status: status,
-                      error: toErrorInfo(error),
-                    });
-                  }
-
-                  // Minimal IValueWithChangeEvent implementation - the shape
-                  // monaco.lsp transports expose via their `state` member.
-                  function createTransportState(initial) {
-                    const listeners = new Set();
-                    let current = initial;
-                    return {
-                      get value() { return current; },
-                      set value(next) {
-                        current = next;
-                        for (const listener of Array.from(listeners)) {
-                          try { listener(next); } catch (_) {}
-                        }
-                      },
-                      get onChange() {
-                        return (listener) => {
-                          listeners.add(listener);
-                          return { dispose: () => listeners.delete(listener) };
-                        };
-                      },
-                    };
-                  }
-
-                  // IMessageTransport whose wire is the Flutter bridge:
-                  // client->server messages surface as 'lspMessage' events;
-                  // server->client messages arrive via deliverServerMessage.
-                  function createBridgedTransport(id) {
-                    let listener = null;
-                    let closed = false;
-                    const state = createTransportState({ state: 'open' });
-                    return {
-                      state: state,
-                      send(message) {
-                        if (closed) {
-                          return Promise.reject(
-                            new Error('Bridged LSP transport is closed: ' + id));
-                        }
-                        postMessageToFlutter({
-                          event: 'lspMessage',
-                          connectionId: id,
-                          message: message,
-                        });
-                        return Promise.resolve();
-                      },
-                      setListener(next) { listener = next; },
-                      deliver(message) {
-                        if (!closed && listener) listener(message);
-                      },
-                      close(error) {
-                        if (closed) return;
-                        closed = true;
-                        state.value = { state: 'closed', error: error };
-                      },
-                      dispose() { this.close(undefined); },
-                      toString() {
-                        return 'FlutterMonacoBridgedTransport(' + id + ')';
-                      },
-                    };
-                  }
-
-                  // Monaco 0.55.1's MonacoLspClient exposes no dispose() and
-                  // discards the feature DisposableStore it creates in its
-                  // constructor, so a stock client can never unregister its
-                  // providers. This wrapper captures the store and reaches
-                  // into the private fields (_initPromise, _bridge,
-                  // _capabilitiesRegistry, _connection) - all verified against
-                  // the 0.55.1 sources. Kept in this one block so future
-                  // Monaco upgrades have a single place to re-verify.
-                  let FlutterLspClientClass = null;
-                  function getClientClass() {
-                    if (FlutterLspClientClass) return FlutterLspClientClass;
-                    FlutterLspClientClass = class extends monaco.lsp.MonacoLspClient {
-                      // Called from the base constructor, before subclass
-                      // field initializers run - stash on a lazily created
-                      // property instead of a declared field (a declared
-                      // field initializer would overwrite this afterwards).
-                      createFeatures() {
-                        const store = super.createFeatures();
-                        if (!this.__flutterFeatureStores) {
-                          this.__flutterFeatureStores = [];
-                        }
-                        this.__flutterFeatureStores.push(store);
-                        return store;
-                      }
-                      get initialized() { return this._initPromise; }
-                      get lspConnection() { return this._connection; }
-                      dispose() {
-                        for (const store of this.__flutterFeatureStores || []) {
-                          try { store.dispose(); } catch (_) {}
-                        }
-                        this.__flutterFeatureStores = [];
-                        try {
-                          this._bridge && this._bridge.dispose && this._bridge.dispose();
-                        } catch (_) {}
-                        try {
-                          this._capabilitiesRegistry &&
-                            this._capabilitiesRegistry.dispose &&
-                            this._capabilitiesRegistry.dispose();
-                        } catch (_) {}
-                      }
-                    };
-                    return FlutterLspClientClass;
-                  }
-
-                  async function buildTransport(id, payload) {
-                    switch (payload && payload.kind) {
-                      case 'webSocket':
-                        // Rejected by the browser when the URL is not in the
-                        // page's connect-src (see allowedConnectSources).
-                        return await monaco.lsp.WebSocketTransport.connectTo({
-                          address: String(payload.url),
-                        });
-                      case 'bridged':
-                        return createBridgedTransport(id);
-                      case 'custom': {
-                        const registry = window.flutterMonacoLspTransports || {};
-                        const factory = registry[payload.factoryName];
-                        if (typeof factory !== 'function') {
-                          throw new Error(
-                            'Unknown LSP transport factory "' + payload.factoryName +
-                            '". Register it on window.flutterMonacoLspTransports ' +
-                            'before connecting.');
-                        }
-                        return await factory(payload.config || {});
-                      }
-                      default:
-                        throw new Error(
-                          'Unknown LSP transport kind: ' + (payload && payload.kind));
-                    }
-                  }
-
-                  // Monaco 0.55.1's client leaves several optional
-                  // server->client requests unhandled and answers them with
-                  // a method-not-found error. Some servers treat that error
-                  // as fatal - pyright (vscode-languageserver) dies on the
-                  // unhandled promise rejection when its
-                  // workspace/diagnostic/refresh call is rejected. These are
-                  // void requests per the LSP spec, so acknowledge them with
-                  // a null result instead. Monaco re-pulls diagnostics and
-                  // tokens on edits, so ignoring the hint is safe.
-                  function registerBenignRefreshHandlers(entry) {
-                    let channel = null;
-                    try {
-                      channel = entry.client.lspConnection &&
-                        entry.client.lspConnection.connection;
-                    } catch (_) {}
-                    if (!channel ||
-                        typeof channel.registerRequestHandler !== 'function') {
-                      return;
-                    }
-                    const passthrough = {
-                      serializeToJson: (v) => (typeof v === 'undefined' ? null : v),
-                      deserializeFromJson: (v) => ({ hasErrors: false, value: v }),
-                    };
-                    const methods = [
-                      'workspace/diagnostic/refresh',
-                      'workspace/semanticTokens/refresh',
-                      'workspace/codeLens/refresh',
-                      'workspace/inlineValue/refresh',
-                      'workspace/foldingRange/refresh',
-                      'workspace/inlayHint/refresh',
-                    ];
-                    for (const method of methods) {
-                      try {
-                        channel.registerRequestHandler({
-                          method: method,
-                          paramsSerializer: passthrough,
-                          resultSerializer: passthrough,
-                          errorSerializer: passthrough,
-                          isOptional: true,
-                        }, async () => ({ ok: null }));
-                      } catch (_) {
-                        // Monaco already handles this method - keep its
-                        // handler.
-                      }
-                    }
-                  }
-
-                  function watchTransportClose(entry) {
-                    const transport = entry.transport;
-                    try {
-                      if (transport && transport.state &&
-                          typeof transport.state.onChange === 'function') {
-                        entry.stateWatch = transport.state.onChange((s) => {
-                          if (s && s.state === 'closed' && !entry.disposed) {
-                            postStatus(entry.id, 'closed', s.error);
-                          }
-                        });
-                      }
-                    } catch (_) {}
-                  }
-
-                  function disposeEntry(entry) {
-                    if (entry.disposed) return;
-                    entry.disposed = true;
-                    try {
-                      entry.stateWatch && entry.stateWatch.dispose &&
-                        entry.stateWatch.dispose();
-                    } catch (_) {}
-                    try {
-                      entry.client && entry.client.dispose && entry.client.dispose();
-                    } catch (_) {}
-                    try {
-                      entry.transport && entry.transport.dispose &&
-                        entry.transport.dispose();
-                    } catch (_) {}
-                    connections.delete(entry.id);
-                    if (connections.size === 0) {
-                      // Monaco's LSP diagnostics all publish under the fixed
-                      // marker owner 'lsp'. Clear them when the last
-                      // connection goes away so stale squiggles don't outlive
-                      // their server.
-                      try {
-                        for (const model of monaco.editor.getModels()) {
-                          monaco.editor.setModelMarkers(model, 'lsp', []);
-                        }
-                      } catch (_) {}
-                    }
-                  }
-
-                  window.flutterMonaco.lsp = {
-                    isAvailable: () => lspAvailable(),
-
-                    async connect(id, transportPayload) {
-                      if (!lspAvailable()) {
-                        throw new Error(
-                          'monaco.lsp is not available in this Monaco build.');
-                      }
-                      if (connections.has(id)) {
-                        throw new Error('LSP connection already exists: ' + id);
-                      }
-                      const entry = {
-                        id: id,
-                        client: null,
-                        transport: null,
-                        bridged: null,
-                        stateWatch: null,
-                        disposed: false,
-                      };
-                      // Register synchronously (before any await) so bridged
-                      // server->client messages can be delivered while the
-                      // initialize handshake is still in flight - the
-                      // handshake cannot complete without them.
-                      connections.set(id, entry);
-                      try {
-                        if (transportPayload && transportPayload.kind === 'bridged') {
-                          entry.bridged = createBridgedTransport(id);
-                          entry.transport = entry.bridged;
-                        } else {
-                          entry.transport = await buildTransport(id, transportPayload);
-                        }
-                        if (!entry.transport ||
-                            typeof entry.transport.send !== 'function' ||
-                            typeof entry.transport.setListener !== 'function') {
-                          throw new Error(
-                            'LSP transport for "' + id +
-                            '" does not implement IMessageTransport ' +
-                            '(state/send/setListener).');
-                        }
-                        const ClientClass = getClientClass();
-                        entry.client = new ClientClass(entry.transport);
-                        registerBenignRefreshHandlers(entry);
-                        watchTransportClose(entry);
-                        await entry.client.initialized;
-                        if (entry.disposed) {
-                          throw new Error(
-                            'LSP connection "' + id +
-                            '" was disposed during initialization.');
-                        }
-                        postStatus(id, 'open', null);
-                        return true;
-                      } catch (e) {
-                        disposeEntry(entry);
-                        throw e;
-                      }
-                    },
-
-                    disconnect(id) {
-                      const entry = connections.get(id);
-                      if (!entry) return false;
-                      disposeEntry(entry);
-                      postStatus(id, 'closed', null);
-                      return true;
-                    },
-
-                    disconnectAll() {
-                      for (const entry of Array.from(connections.values())) {
-                        disposeEntry(entry);
-                        postStatus(entry.id, 'closed', null);
-                      }
-                      return true;
-                    },
-
-                    // Dart -> JS delivery for bridged transports
-                    // (server -> client direction).
-                    deliverServerMessage(id, message) {
-                      const entry = connections.get(id);
-                      if (!entry || !entry.bridged) return false;
-                      entry.bridged.deliver(message);
-                      return true;
-                    },
-
-                    listConnections() {
-                      return Array.from(connections.keys());
-                    },
-
-                    // EXPERIMENTAL escape hatch for non-standard server
-                    // extensions. Depends on Monaco-internal channel plumbing
-                    // (verified against 0.55.1); may break on future Monaco
-                    // upgrades, in which case it throws a descriptive error.
-                    async sendRequest(id, method, params) {
-                      const entry = connections.get(id);
-                      if (!entry || !entry.client) {
-                        throw new Error('No such LSP connection: ' + id);
-                      }
-                      const lspConnection = entry.client.lspConnection;
-                      const channel = lspConnection && lspConnection.connection;
-                      if (!channel || typeof channel.request !== 'function') {
-                        throw new Error(
-                          'Generic LSP requests are not supported by this ' +
-                          'Monaco build.');
-                      }
-                      const passthrough = {
-                        serializeToJson: (v) => (typeof v === 'undefined' ? null : v),
-                        deserializeFromJson: (v) => ({ hasErrors: false, value: v }),
-                      };
-                      return await channel.request({
-                        method: String(method),
-                        paramsSerializer: passthrough,
-                        resultSerializer: passthrough,
-                        errorSerializer: passthrough,
-                        isOptional: false,
-                      }, params);
-                    },
-
-                    async sendNotification(id, method, params) {
-                      const entry = connections.get(id);
-                      if (!entry || !entry.client) {
-                        throw new Error('No such LSP connection: ' + id);
-                      }
-                      const lspConnection = entry.client.lspConnection;
-                      const channel = lspConnection && lspConnection.connection;
-                      if (!channel || typeof channel.notify !== 'function') {
-                        throw new Error(
-                          'Generic LSP notifications are not supported by ' +
-                          'this Monaco build.');
-                      }
-                      const passthrough = {
-                        serializeToJson: (v) => (typeof v === 'undefined' ? null : v),
-                      };
-                      channel.notify({
-                        method: String(method),
-                        paramsSerializer: passthrough,
-                      }, params);
-                      return true;
-                    },
-                  };
-                })();
-              })();
-
-              postMessageToFlutter({ event: 'onEditorReady' });
-              console.log('[Monaco] Editor is ready and the Flutter bridge is installed.');
-            });
-
-            monaco.editor.create(document.getElementById('editor-container'), {
-              value: '// Monaco Editor is ready',
-              language: 'markdown',
-              theme: 'vs-dark',
-              automaticLayout: true,
-              wordWrap: 'on',
-              padding: { top: 10 },
-              minimap: { enabled: false }
-            });
-          },
-          function (error) {
-            console.error('[Monaco] FATAL: require() failed to load editor.main.js. Error:', error);
-            ${isWeb ? "if (window.flutterMonacoPostMessage) window.flutterMonacoPostMessage({ event: 'error', message: 'Failed to load editor.main: ' + error });" : ''}
-          }
-        );
-      } catch (e) {
-        console.error('[Monaco] FATAL: A critical error occurred trying to call require(). Error:', e);
-      }
-      ${isWeb ? '}; _initMonaco();' : '})();'}
-    </script>
-  </body>
-</html>
-''';
+    return buildMonacoIndexHtml(
+      vsPath: vsPath,
+      bridgeBase: bridgeBasePath,
+      isWindows: isWindows,
+      isIosOrMacOS: isIosOrMacOS,
+      isWeb: isWeb,
+      messageToken: messageToken,
+      customCss: customCss,
+      allowCdnFonts: allowCdnFonts,
+      allowedConnectSources: allowedConnectSources,
+    );
   }
 }
